+++ hashcathelper/subcommands/analytics.py
@@ -75,7 +75,7 @@ args.append(argument(
     default=None,
     help="""
 URL to a Neo4j database containing BloodHound data. Format:
-bolt[s]://<user>:<password>@<host>:<port>"""
+bolt[s]://<user>:<password>@<host>[:<port>]"""
 ))